requirements.txt
Django==4.1.2
djangorestframework-simplejwt
django-cors-headers  
djoser